requirements.txt
bcrypt==3.2.2
blinker==1.4
git+https://github.com/helloflask/bootstrap-flask
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
dominate==2.6.0
et-xmlfile==1.1.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-DebugToolbar==0.13.1
Flask-DebugToolbar-LineProfilerPanel==1.0.2
Flask-WTF==1.0.1
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
gprof2dot==2021.2.21
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
line-profiler==3.5.1
MarkupSafe==2.1.1
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.2.0
openpyxl==3.0.9
pandas==1.4.2
paramiko==2.10.4
pprofile==2.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDrive==1.3.1
PyNaCl==1.5.0
pyparsing==3.0.8
pysftp==0.2.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
snakeviz==2.1.1
tornado==6.1
uritemplate==4.1.1
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
